requirements.txt
Django
djangorestframework
environs
gunicorn
whitenoise
dj-database-url